Cleaning up C# code.

+++ SDK/Runner/Workspace/FileSystemExtensions.cs
@@ -31,32 +31,32 @@ namespace PixelVision8.Runner.Workspace
 {
     public static class FileSystemExtensions
     {
-//        public static Stream Open(this File file, FileAccess access)
-//        {
-//            return file.FileSystem.OpenFile(file.Path, access);
-//        }
-//
-//        public static void Delete(this FileSystemEntity entity)
-//        {
-//            entity.FileSystem.Delete(entity.Path);
-//        }
+        //        public static Stream Open(this File file, FileAccess access)
+        //        {
+        //            return file.FileSystem.OpenFile(file.Path, access);
+        //        }
+        //
+        //        public static void Delete(this FileSystemEntity entity)
+        //        {
+        //            entity.FileSystem.Delete(entity.Path);
+        //        }
 
-//        public static ICollection<FileSystemPath> GetEntityPaths(this Directory directory)
-//        {
-//            return directory.FileSystem.GetEntities(directory.Path);
-//        }
+        //        public static ICollection<FileSystemPath> GetEntityPaths(this Directory directory)
+        //        {
+        //            return directory.FileSystem.GetEntities(directory.Path);
+        //        }
 
-//        public static ICollection<FileSystemEntity> GetEntities(this Directory directory)
-//        {
-//            var paths = directory.GetEntityPaths();
-//            return new EnumerableCollection<FileSystemEntity>(
-//                paths.Select(p => FileSystemEntity.Create(directory.FileSystem, p)), paths.Count);
-//        }
+        //        public static ICollection<FileSystemEntity> GetEntities(this Directory directory)
+        //        {
+        //            var paths = directory.GetEntityPaths();
+        //            return new EnumerableCollection<FileSystemEntity>(
+        //                paths.Select(p => FileSystemEntity.Create(directory.FileSystem, p)), paths.Count);
+        //        }
 
         public static IEnumerable<WorkspacePath> GetEntitiesRecursive(this IFileSystem fileSystem, WorkspacePath path)
         {
-            if (!path.IsDirectory)
-                throw new ArgumentException("The specified path is not a directory.");
+            if (!path.IsDirectory) throw new ArgumentException("The specified path is not a directory.");
+
             foreach (var entity in fileSystem.GetEntities(path))
             {
                 yield return entity;
@@ -68,14 +68,13 @@ namespace PixelVision8.Runner.Workspace
 
         public static void CreateDirectoryRecursive(this IFileSystem fileSystem, WorkspacePath path)
         {
-            if (!path.IsDirectory)
-                throw new ArgumentException("The specified path is not a directory.");
+            if (!path.IsDirectory) throw new ArgumentException("The specified path is not a directory.");
+
             var currentDirectoryPath = WorkspacePath.Root;
             foreach (var dirName in path.GetDirectorySegments())
             {
                 currentDirectoryPath = currentDirectoryPath.AppendDirectory(dirName);
-                if (!fileSystem.Exists(currentDirectoryPath))
-                    fileSystem.CreateDirectory(currentDirectoryPath);
+                if (!fileSystem.Exists(currentDirectoryPath)) fileSystem.CreateDirectory(currentDirectoryPath);
             }
         }
 
@@ -104,8 +103,8 @@ namespace PixelVision8.Runner.Workspace
             }
             else
             {
-                if (!destinationPath.IsRoot)
-                    destinationFileSystem.CreateDirectory(destinationPath);
+                if (!destinationPath.IsRoot) destinationFileSystem.CreateDirectory(destinationPath);
+
                 foreach (var ep in sourceFileSystem.GetEntities(sourcePath))
                 {
                     var destinationEntityPath = ep.IsFile
@@ -115,30 +114,30 @@ namespace PixelVision8.Runner.Workspace
                 }
             }
 
-//            IEntityCopier copier;
-//            if (!EntityCopiers.Registration.TryGetSupported(sourceFileSystem.GetType(), destinationFileSystem.GetType(),
-//                out copier))
-//                throw new ArgumentException("The specified combination of file-systems is not supported.");
-//            copier.Copy(sourceFileSystem, sourcePath, destinationFileSystem, destinationPath);
+            //            IEntityCopier copier;
+            //            if (!EntityCopiers.Registration.TryGetSupported(sourceFileSystem.GetType(), destinationFileSystem.GetType(),
+            //                out copier))
+            //                throw new ArgumentException("The specified combination of file-systems is not supported.");
+            //            copier.Copy(sourceFileSystem, sourcePath, destinationFileSystem, destinationPath);
         }
 
-//        public static void CopyTo(this FileSystemEntity entity, IFileSystem destinationFileSystem,
-//            FileSystemPath destinationPath)
-//        {
-//            entity.FileSystem.Copy(entity.Path, destinationFileSystem, destinationPath);
-//        }
-//
-//        public static void CopyTo(this Directory source, Directory destination)
-//        {
-//            source.FileSystem.Copy(source.Path, destination.FileSystem,
-//                destination.Path.AppendDirectory(source.Path.EntityName));
-//        }
-//
-//        public static void CopyTo(this File source, Directory destination)
-//        {
-//            source.FileSystem.Copy(source.Path, destination.FileSystem,
-//                destination.Path.AppendFile(source.Path.EntityName));
-//        }
+        //        public static void CopyTo(this FileSystemEntity entity, IFileSystem destinationFileSystem,
+        //            FileSystemPath destinationPath)
+        //        {
+        //            entity.FileSystem.Copy(entity.Path, destinationFileSystem, destinationPath);
+        //        }
+        //
+        //        public static void CopyTo(this Directory source, Directory destination)
+        //        {
+        //            source.FileSystem.Copy(source.Path, destination.FileSystem,
+        //                destination.Path.AppendDirectory(source.Path.EntityName));
+        //        }
+        //
+        //        public static void CopyTo(this File source, Directory destination)
+        //        {
+        //            source.FileSystem.Copy(source.Path, destination.FileSystem,
+        //                destination.Path.AppendFile(source.Path.EntityName));
+        //        }
 
         #endregion
 
@@ -180,34 +179,33 @@ namespace PixelVision8.Runner.Workspace
                     Move(sourceFileSystem, ep, destinationFileSystem, destinationEntityPath);
                 }
 
-                if (!sourcePath.IsRoot)
-                    sourceFileSystem.Delete(sourcePath);
+                if (!sourcePath.IsRoot) sourceFileSystem.Delete(sourcePath);
             }
 
-//            IEntityMover mover;
-//            if (!EntityMovers.Registration.TryGetSupported(sourceFileSystem.GetType(), destinationFileSystem.GetType(),
-//                out mover))
-//                throw new ArgumentException("The specified combination of file-systems is not supported.");
-//            mover.Move(sourceFileSystem, sourcePath, destinationFileSystem, destinationPath);
+            //            IEntityMover mover;
+            //            if (!EntityMovers.Registration.TryGetSupported(sourceFileSystem.GetType(), destinationFileSystem.GetType(),
+            //                out mover))
+            //                throw new ArgumentException("The specified combination of file-systems is not supported.");
+            //            mover.Move(sourceFileSystem, sourcePath, destinationFileSystem, destinationPath);
         }
 
-//        public static void MoveTo(this FileSystemEntity entity, IFileSystem destinationFileSystem,
-//            FileSystemPath destinationPath)
-//        {
-//            entity.FileSystem.Move(entity.Path, destinationFileSystem, destinationPath);
-//        }
+        //        public static void MoveTo(this FileSystemEntity entity, IFileSystem destinationFileSystem,
+        //            FileSystemPath destinationPath)
+        //        {
+        //            entity.FileSystem.Move(entity.Path, destinationFileSystem, destinationPath);
+        //        }
 
-//        public static void MoveTo(this Directory source, Directory destination)
-//        {
-//            source.FileSystem.Move(source.Path, destination.FileSystem,
-//                destination.Path.AppendDirectory(source.Path.EntityName));
-//        }
+        //        public static void MoveTo(this Directory source, Directory destination)
+        //        {
+        //            source.FileSystem.Move(source.Path, destination.FileSystem,
+        //                destination.Path.AppendDirectory(source.Path.EntityName));
+        //        }
 
-//        public static void MoveTo(this File source, Directory destination)
-//        {
-//            source.FileSystem.Move(source.Path, destination.FileSystem,
-//                destination.Path.AppendFile(source.Path.EntityName));
-//        }
+        //        public static void MoveTo(this File source, Directory destination)
+        //        {
+        //            source.FileSystem.Move(source.Path, destination.FileSystem,
+        //                destination.Path.AppendFile(source.Path.EntityName));
+        //        }
 
         #endregion
     }
